+++ tools/mhmake/src/mhmakefileparser.h
@@ -100,7 +100,7 @@ protected:
 #else
   char                **m_pEnv;            // New environment in case the makefile exports variables
 #endif
-  int                   m_EnvLen;          // Current length of m_pEnv
+  size_t                m_EnvLen;          // Current length of m_pEnv
 
   autodeps_t m_AutoDeps;
   set< const fileinfo* , less_fileinfo > m_Targets; // List of targets that are build by this makefile
@@ -363,7 +363,7 @@ public:
   mh_pid_t EchoCommand(const string &Params) const;
   string SearchCommand(const string &Command, const string &Extension="") const;
   const string &GetPythonExe() const;
-  int SearchPath(const char *szCommand, const char *pExt, int Len, char *szFullCommand,char **pFilePart) const;
+  int SearchPath(const char *szCommand, const char *pExt, size_t Len, char *szFullCommand,char **pFilePart) const;
   mh_pid_t OsExeCommand(const string &Command, const string &Params, bool IgnoreError, string *pOutput) const;
 };
